region 조합 key 로 변경

+++ hibernate-redis/src/main/java/org/hibernate/cache/redis/RedisRegionFactory.java
@@ -1,9 +1,10 @@
 package org.hibernate.cache.redis;
 
-import lombok.extern.slf4j.Slf4j;
 import org.hibernate.cache.CacheException;
 import org.hibernate.cache.redis.util.RedisTool;
 import org.hibernate.cfg.Settings;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.Properties;
 
@@ -13,8 +14,11 @@ import java.util.Properties;
  * @author sunghyouk.bae@gmail.com
  * @since 13. 4. 6. 오전 12:41
  */
-@Slf4j
 public class RedisRegionFactory extends AbstractRedisRegionFactory {
+
+    private static final Logger log = LoggerFactory.getLogger(RedisRegionFactory.class);
+    private static final boolean isTranceEnabled = log.isTraceEnabled();
+    private static final boolean isDebugEnabled = log.isDebugEnabled();
 
     private RedisClient redis;
 
@@ -24,13 +28,15 @@ public class RedisRegionFactory extends AbstractRedisRegionFactory {
 
     @Override
     public void start(Settings settings, Properties properties) throws CacheException {
-        if (log.isDebugEnabled())
-            log.debug("Start region factory");
+        if (isDebugEnabled)
+            log.debug("Starting region factory...");
 
         this.settings = settings;
 
         try {
             this.redis = RedisTool.createRedisClient(props);
+            if (isDebugEnabled)
+                log.debug("Start region factory!!!");
         } catch (Exception e) {
             throw new CacheException(e);
         }
@@ -38,12 +44,15 @@ public class RedisRegionFactory extends AbstractRedisRegionFactory {
 
     @Override
     public void stop() {
-        if (log.isDebugEnabled())
-            log.debug("Stop regoin factory");
+        if (redis == null) return;
+        if (isDebugEnabled)
+            log.debug("Stop regoin factory...");
 
         try {
             redis.flushDb();
             redis = null;
+            if (isDebugEnabled)
+                log.debug("Stop region factory!!!");
         } catch (Exception e) {
             log.error("redis region factory fail to stop.", e);
         }
